# faster traversable sorting with the help of vector-algorithms