--- a/sort-traversable.cabal
+++ b/sort-traversable.cabal
@@ -47,6 +47,7 @@ library
     Data.Traversable.Sort.PairingHeap
     Data.Traversable.Sort.PairingHeap.BasicNat
     Data.Traversable.Sort.PairingHeap.IndexedPairingHeap
+    Data.Traversable.Sort.Vector
 
   -- Modules included in this library but not exported.
   -- other-modules:
@@ -65,6 +66,9 @@ library
 
   -- Other library packages from which modules are imported.
   build-depends:       base >=4.8 && <4.10
+                     , vector
+                     , vector-algorithms
+                     , mtl
 
   -- Directories containing source files.
   hs-source-dirs: src, benchmarks